Generate instanceTest for each interfaces (that have instanceTests)
