+++ wasm/src/main/scala/ir2wasm/HelperFunctions.scala
@@ -3,6 +3,8 @@ package wasm.ir2wasm
 import org.scalajs.ir.{Trees => IRTrees}
 import org.scalajs.ir.{Types => IRTypes}
 import org.scalajs.ir.{Names => IRNames}
+import org.scalajs.linker.standard.LinkedClass
+import org.scalajs.ir.ClassKind
 
 import wasm.wasm4s._
 import wasm.wasm4s.WasmContext._
@@ -682,6 +684,100 @@ object HelperFunctions {
       instrs += CALL(FuncIdx(WasmFunctionName.getClassOf))
     }
 
+    fctx.buildAndAddToContext()
+  }
+
+  /** Generate type inclusion test for interfaces `isInstanceOf[<interface>]` will be compiled to
+    * the CALL of the generated this function.
+    *
+    * TODO: Efficient type inclusion test Current implementation walk through the itables of the
+    * Object which takes O(N) (where N = number of interfaces the expr implements) See:
+    * https://github.com/tanishiking/scala-wasm/issues/27#issuecomment-2008252049
+    */
+  def genInstanceTest(clazz: LinkedClass)(implicit ctx: WasmContext): Unit = {
+    import WasmImmediate._
+    assert(clazz.kind == ClassKind.Interface)
+
+    val fctx = WasmFunctionContext(
+      Names.WasmFunctionName.instanceTest(clazz.name.name),
+      List("expr" -> WasmAnyRef),
+      List(WasmInt32)
+    )
+    val List(exprParam) = fctx.paramIndices
+
+    import fctx.instrs
+
+    val found = fctx.addLocal(fctx.genSyntheticLocalName(), Types.WasmInt32)
+    val cnt = fctx.addLocal(fctx.genSyntheticLocalName(), Types.WasmInt32)
+    val len = fctx.addLocal(fctx.genSyntheticLocalName(), Types.WasmInt32)
+    val itables = fctx.addLocal(
+      fctx.genSyntheticLocalName(),
+      Types.WasmRefNullType(Types.WasmHeapType.Type(WasmArrayType.itables.name))
+    )
+
+    instrs += I32_CONST(I32(0))
+    instrs += LOCAL_SET(found)
+
+    fctx.block() { testFail =>
+      // if expr is not an instance of Object, return false
+      instrs += LOCAL_GET(exprParam)
+      instrs += REF_TEST(HeapType(Types.WasmHeapType.ObjectType))
+      instrs += I32_CONST(I32(1))
+      instrs += I32_XOR
+      instrs += BR_IF(testFail)
+
+      // if the itables is null (no interfaces are implemented)
+      instrs += LOCAL_GET(exprParam)
+      instrs += REF_CAST(HeapType(Types.WasmHeapType.ObjectType))
+      instrs += STRUCT_GET(TypeIdx(Types.WasmHeapType.ObjectType.typ), StructFieldIdx(1))
+      instrs += LOCAL_TEE(itables)
+      instrs += REF_IS_NULL
+      instrs += BR_IF(testFail)
+
+      // found := 0
+      // len := length(itables)
+      // loop $loopLabel {
+      //   if (cnt < len) {
+      //     if (itables(cnt) is instance of testClassName's itable) {
+      //       found := 1
+      //     } else {
+      //       cnt := cnt + 1
+      //       br $loopLabel
+      //   }
+      // }
+      // return found
+      instrs += I32_CONST(I32(0))
+      instrs += LOCAL_SET(cnt)
+      // len := length(itables)
+      instrs += LOCAL_GET(itables)
+      instrs += ARRAY_LEN
+      instrs += LOCAL_SET(len)
+
+      fctx.loop() { loopLabel =>
+        instrs += LOCAL_GET(cnt)
+        instrs += LOCAL_GET(len)
+        instrs += I32_LT_U
+        fctx.ifThen() {
+          instrs += LOCAL_GET(itables)
+          instrs += LOCAL_GET(cnt)
+          instrs += ARRAY_GET(TypeIdx(WasmArrayType.itables.name))
+          instrs += REF_TEST(
+            HeapType(Types.WasmHeapType.Type(WasmTypeName.WasmITableTypeName(clazz.name.name)))
+          )
+          fctx.ifThenElse() {
+            instrs += I32_CONST(I32(1))
+            instrs += LOCAL_SET(found)
+          } {
+            instrs += LOCAL_GET(cnt)
+            instrs += I32_CONST(I32(1))
+            instrs += I32_ADD
+            instrs += LOCAL_SET(cnt)
+            instrs += BR(loopLabel)
+          }
+        }
+      }
+    }
+    instrs += LOCAL_GET(found)
     fctx.buildAndAddToContext()
   }
